requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pdfplumber==0.11.4
chromadb==0.5.23
langchain==0.3.12
langchain-community==0.3.12
google-genai==1.10.0
langchain-groq==0.2.3
sentence-transformers==3.3.1
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0